Fix: "Scale to Fit (Letterbox)" crops off the edges
Rather than deciding based on the screen dimensions as is done currently (which is flawed), I am now calculating the scale using the minimum to get a side to the edge for scale to fit letterbox and the maximum for scale to fit fill. This means that letterboxing can occur on both the x and the y axis.