Start in fullscreen broken on macOS target
http://community.stencyl.com/index.php?issue=1639.0

The last event we receive when going to fullscreen on macos is a window maximized event (see SDL_cocoawindow.m), and this happens before the correcting window event listeners have been registered on the engine.

The changes were tested on Windows as well, and there seemed to be no difference in behavior without the extra resize/fullscreen event dispatch from Universal, so those were removed.