Fix maskLayer math.
The previous code was wrong, but in most cases (most cases being a scale value > 1) the issue wasn't noticeable since it would only cause the masking area to grow away from the center of the screen. On the other hand, if the basic dimensions of the game were larger than the screen, and we actually needed to scale the game -down- to enter fullscreen, that would cause the maskLayer to scale down with it, exposing parts of the screen.