Fix event cleanup logic when killing actors.
This has been an issue since Nov 2013 -- 516f169125ad0bb53a4afc63a234bdf0c7e35e3d. The removeAllListeners function does what it's supposed to, so this was never noticed. It seems that, at worst, this may have led to some wastefulness over many consecutive lifetimes of a single actor.

It was noticed now because it causes a crash on hashlink.