[input] Make Control.buttons and _joyControlMap both use Strings.
Making Control.buttons and _joyControlMap both use the same type makes them more similar to Control.keys and _keyControlMap. This is the first step toward abstracting the two types of input out to a common class.

The only time JoystickButton is actually used is in controlStateUpdated. That should happen infrequently enough that there's no reason to avoid a map lookup there.