Return -1 instead of 0 from getIDForScene for scenes that don't exist
0 isn't a good return value here for missing scenes, since 0 is a valid
scene ID.

This function is only used by design mode in the "get scene by name"
block, with the effect that asking for a scene that doesn't exist will
return the scene with ID 0, if it exists, or null otherwise. After this
change, scenes that don't exist will always return null, making the
"get scene by name" block suitable for checking whether or not a scene
exists.