Don't move joystick center on release
Center doesn't need to be set here. For a static joystick, center should
only be changed by the user. For a relative joystick, center is set at
every press.

By default, outerRadius is half of the size of outerImage, so unless a
custom outerRadius is set, center remains the same, hiding the problem.