Add images to batched render, allow them to be attached to actors
To track which BitmapDatas have been uploaded to a GPU texture, a
WeakMap is used. Javascript doesn't have haxe.ds.WeakMap, but it
does have js.lib.WeakMap, with different type parameters and no
map iteration or clearing.

To keep from changing a lot of the API surface, BitmapWrapper is
repurposed for use with Tiles, but this requires it to be passed
a BitmapData instead of a Bitmap.

It's a bit awkward for Layer to support image placement both in front
of and behind the main sublayers. Since images are tiles they need to
be placed in Tilemaps, so it's impossible to get an implementation
identical to the Bitmap one. In particular, the bitmap-based images can
be moved around freely, even between sublayers. The solution here
(having separate front/back sublayers for images) doesn't allow this,
but having images partitioned like this seems like a better solution,
and I'm tempted to make normal Bitmap-based images work the same way.

The restriction on using filters is enforced for images now as well.