Change gci timeout function to work better on Flash
It seems that timers and socket connections don't work right away. This
appears to be related to the streaming nature of SWF content playback.
Until the main SWF is fully loaded, the timer won't move again and the
socket won't attempt to connect.

Instead of setting the startTime directly in create (which runs nearly
immediately), set startTime when the tryTimeout function first runs
(which could be many seconds later, especially if a debugger is
running).