Treat the engine as an engine extension
The toolset now treats the engine itself the same way it treats other
engine extensions. This is how the newly-defined blocks.xml file is
read in.

This also allows the engine to be vendored on a per-project basis in the
same manner that other engine extensions are: by copying it into the
engine-extensions folder of the project.