Catch and log errors in custom event triggers
Flash and other targets worked differently here which complicated this
code a little. Here are the results of using reflection on different
targets:

Reflect.hasField(script, "functionName")
 if a function exists:        true on flash, false on other targets
 if a function doesn't exist: false

Reflect.field(script, "functionName")
 if a function exists:        function object
 if a function doesn't exist: null

Reflect.callMethod(script, Reflect.field(script, "functionName"), [])
 if a function exists:        return value of function
 if a function doesn't exist:
   flash: Exception: Error #1009: Cannot access a property or method of a null object reference.
   html5: Exception: Cannot read properties of undefined (reading 'apply')
   native: Exception: Invalid function null

Based on this, it seems the conditional compilation can be removed here
by just using Reflect.field to check the existence of a function.

Additionally, the way that errors are handled has been made more
consistent.

The previous behavior on non-flash targets was to swallow errors,
assuming it to be an error in reflection, and to try using
forwardMessage instead (presumably never matching a real message name).

The previous behavior on flash was to just call the function directly
with no try-catch, which would lead to an unhandled error.

Now, the function call is always wrapped in a try-catch for error
logging.

Finally, if using call2, which targets a specific behavior, a warning is
logged if that behavior doesn't exist or is not attached.