Simplify background code
Removed some completely unused or needless fields
- currFrame, currTime in ImageBackground
- curStep, speed in ScrollingBitmap
- bitmap, cacheWidth, cacheHeight in BackgroundLayer

Reimplemented ScrollingBitmap so that it works with any image size.
Previously, it was hardcoded to 9 copies of the background texture,
so the background was repeated in advance to be large enough.

Now any image can be used, and enough bitmaps will be created to fill
the screen.

This also allowed the removal of a few more fields:
- img and repeated in ImageBackground
- cacheIndex in BackgroundLayer